refactor: remove circular dependency between api_resilience.ts and metrics_collector.ts
- Created resilience_types.ts with shared types and interfaces
- Moved CircuitBreakerState and CircuitBreakerMetrics to shared types
- Added MetricsCollectorInterface for dependency injection
- Updated ResilientApiWrapper to accept optional metrics collector
- Modified constructor to use injected metrics collector instead of direct import
- Updated test files to import types from correct locations

Resolves circular dependency issue where api_resilience imported from
metrics_collector which also imported from api_resilience.
